data-processing/machine-learning/requirements.txt
# CityFlow Machine Learning Pipeline - Dependencies

# Core ML Libraries
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
tensorflow==2.18.0  # Updated for Python 3.12 compatibility

# Data Processing
pandas==2.1.4
numpy==1.26.2
pyarrow==14.0.1

# Feature Engineering
feature-engine==1.6.2

# MLflow for Experiment Tracking
mlflow==2.9.2
mlflow-skinny==2.9.2

# Model Serving
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Kafka Integration
confluent-kafka==2.3.0
avro-python3==1.10.2  # Latest version compatible with Python 3.12
fastavro==1.9.3

# Database Connectivity
psycopg2-binary==2.9.9
pymongo==4.6.1
redis==5.0.1

# Data Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Time Series
prophet==1.1.5
statsmodels==0.14.1

# Geospatial
geopy==2.4.1
shapely==2.0.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
requests==2.31.0
httpx==0.26.0

# Logging and Monitoring
prometheus-client==0.19.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.23.3
pytest-cov==4.1.0
